+++ mdsuite/file_io/file_io_dict.py
@@ -9,6 +9,6 @@ from mdsuite.file_io.lammps_trajectory_files import LAMMPSTrajectoryFile
 from mdsuite.file_io.lammps_flux_files import LAMMPSFluxFile
 
 dict_file_io = {
-    'lammpstraj': (LAMMPSTrajectoryFile, 'traj'),
-    'lammpsflux': (LAMMPSFluxFile, 'flux'),
+    'lammps_traj': (LAMMPSTrajectoryFile, 'traj'),
+    'lammps_flux': (LAMMPSFluxFile, 'flux'),
 }
